Add support for dotfiles to `boxen::personal`
For a while now many people have been managing their dotfiles either
outside of Boxen or using their own custom modules. While not a bad
thing, it definitely adds some pain to using Boxen.

To address this, I've added a new `boxen::dotfiles` resource. In a bid
to keep this integration as agnostic as possible, you only need to
define two parameters, `repository` and `install_command`. The
repository is the git URL you would like to fetch the settings from and
the `install_command` is what will trigger the setup of your dotfiles in
whatever manner you desire. Here is an example using hiera:

```yaml
boxen::personal::dotfiles:
  repository: https://github.com/example/dotfiles.git
  install_command: './install.sh'
```

The plan here is that we keep the specifics of how you manage your
dotfiles **out of Boxen** and inside of the repository itself.

Closes #141